add run_until_graceful_shutdown method with shutdown hook
